# Improve account deletion performances further