## Benchmark

### Protocol

Before each test, the database is re-created with the same backup from my production instance (~single-user, a few local accounts for different purposes). Media files are **not** copied over.

The development environment does not have ElasticSearch enabled, and the database server runs locally.

The timing values are obtained by running the following code in a Rails console:

```ruby
Benchmark.measure do
  top_account_ids.each do |account_id|
    DeleteAccountService.new.call(Account.find(account_id), skip_activitypub: true)
  end
end
```

`top_account_ids` contains the identifier of the 5 accounts with the most toots known to my instance, they are broken down as follows:

| # | total toots | public toots | DMs | local? | favourites | reported toots |
| - | -- | -- | -- | -- | -- | -- |
| 1 | 105207 | 13044 | 48 | no | 1054 | 1 |
| 2 | 73196 | 66623 | 0 | no | 369 | 0 |
| 3 | 57971 | 19237 | 329 | no | 1921 | 0 |
| 4 | 56710 | 4067 | 1255 | yes | 37959 | 0 |
| 5 | 54403 | 31091 | 56 | no | 3148 | 0 |

The accounts are deleted in that order, and interact between them, so an account being cleared will slightly lower the number of toots (reblogs) to be processed in other accounts.

### Results

| commit | user | system | total | real |
| --- | --- | --- | --- | --- |
| 2334addfff285fc69d616c40ee0fb8a696fcd3f8 (without this PR) | 750.049139s (12min30s) |  55.861878s | 805.911017s (13min26s) | 2624.943940s (**43min45s**) |
| 8c70b012cdfd7a97082300275331c93bd402a7c6 (deletes statuses by batches of 50 instead of individually) | 606.902921s (10min07s) | 41.281127s | 648.184048s (10min48s) | 1549.474583s (**25min50s**) |
| e20b52ee43c4b17f78b807f960781f03a6f98695 (do not precompute values that are only used once) | 596.145949s (9min56s) | 42.301746s | 638.447695s (10min38s) | 1527.635537s (25min27s) |
| fb8ad0f6f8a0d9cb6ffb90fc1d285aac45b54b32 (do not generate redis events for toots older than two weeks) | 479.127842s (7min59s) | 40.115994s | 519.243836s (8min39s) | 1448.805232s (24min09s) |
| eff219a55eb59fa0465a81ecde4c8874a2c2ad4c (filter reported toots a priori) | 476.933216s (7min57s) | 39.721144s | 516.654360s (8min37s) | 1424.850713s (23min45s) |
| fc6a592d0426ed667ff45bc450581358487645a9 (do not process reblogs when cleaning up public timelines) | 477.414900s (7min57s) | 39.666471s | 517.081371s (8min37s) | 1426.631086s (23min47s) |
| d5da93708604efd209e1b55b0fd5610bddbbfe18 (clean up deleted account's feed in one go) | 477.063304s (7min57s) | 37.829388s | 514.892692s (8min35s) | 1413.292442s (23min33s) |
| f57ed408598a2dc6c9ff701c1a9d9efb4069be2e (delete rather than destroy a few more associations, tweak preloading) | 474.929086s (7min55s) | 38.475544s | 513.404630s (8min33s) | 1415.174932s (**23min35s**) |

The first few commits make a very sizable difference, then the benefits aren't that obvious on my example. Still, they make the logic a bit cleaner and I expect it to make a slightly bigger difference on instances with more accounts interacting with the deleted ones.