@ThibG 2020-12-22
In addition, we should be able to do smart things for `favourites` and `bookmarks` but I'll have to read up on how Chewy works to efficiently update the indexes though.