@Gargron 2020-12-22
> In addition, we should be able to do smart things for `favourites` and `bookmarks`

I think you mean `favourites`, `active_relationships` and `passive_relationships` - `bookmarks` have no associated counters. I already have some code for that...