@ThibG 2020-12-22
`bookmarks` has no associated counters but it has ES indexing