@ThibG 2020-12-22
I think I'm about done with this PR, I don't really know where to trim some processing time regarding statuses (except maybe making `unpush_from_home_timelines` and `unpush_from_list_timelines` smarter, but it's tricky).

There's lots of possible improvements in `favourites`, `active_relationships` and `passive_relationships` (and to a lesser extent `bookmarks`) but I understood from your earlier post that you were experimenting with them, so I'm leaving that up to you for the time being.

All in all, here is a breakdown of the time spent deleting each of the accounts with the latest commit of this PR:

| # | statuses | media attachments | notifications | active_relationships | passive_relationships | favourites |
| - | ----------- | ------------------------- | ---------------- | -------------------------- | ----------------------------- | ------------- |
| 1 | 319118.9ms | 4947.4ms | 274.0ms | 54.8ms | 25.0ms | 20304.3ms |
| 2 | 174690.6ms | 12454.8ms | 31.2ms | 23.7ms | 11.8ms | 4233.3ms |
| 3 | 192512.5ms | 8221.9ms | 149.2ms | 20.8ms | 30.5ms | 15070.6ms |
| 4 | 196633.6ms | 17311.3ms | 17.0ms | 2128.6ms | 581.0ms | 292957.6ms |
| 5 | 142705.5ms | 6301.1ms | 25.8ms | 52.6ms | 13.0ms | 4947.7ms |